fix: rename utils to hooks folder
